Update allocated-pids.txt (#180)
Adding Barduino PID's.

Barduino is a development board created by Fablab Barcelona as a tool for it's educational programs. 

ESP32-S3 (ESP32-S3-WROOM-1)

It needs a custom PID to have the opportunity to show it officialy in the CircuitPython website and to create an Arduino core for it. 

It's done under Fablab Barcelona, part of IAAC. 

This is (will be because it's a WIP) the documentation for the board https://fablabbcn-projects.gitlab.io/electronics/barduino-docs/
and this is Fablab Barcelona https://fablabbcn.org/
and this IAAC https://iaac.net/

Thank you!