Update allocated-pids.txt (#204)
Please answer the questions in the README.md of this repo:

Give a short description of the device and its function, Smart meter for home storage battery's using LORA comunications between the smart meter and the reciever.
Tell us what chip you're using, ESP32-S2FH4
Mention why you need a custom PID, For updating the firmware using UF2 bootloader and for usb serial comunication to Victron inverters.
If applicable/available mention your company and a link to the website of the product. None available.