- Give a short description of the device and its function: (#268)
These are RFID card and USB fingerprint reader devices.
- Tell us what chip you're using:
The device are based on the ESP32-S3 chip.
- Mention why you need a custom PID:
The device uses USB communication to connect to machines and requires custom PIDs so that the embedded software can identify which device is connected and manage communication appropriately.
- If applicable/available mention your company and a link to the website of the product:
Neubras, website: https://neubras.com